requirements.txt
torch==2.3.1
torvision==0.18.1
torchaudio==2.3.1
drake
healpy
manipulation
matplotlib
mesh_to_sdf
numpy
omegaconf
open3d
Pillow
plotly
pykdtree
scikit_learn
scipy
torch_scatter
trimesh
yacs
opencv-python